remove admin routes and rename app.js to index.js; update package.json for script changes
