+++ admin/backend/package.json
@@ -4,7 +4,7 @@
   "description": "mern stack",
   "main": "index.js",
   "scripts": {
-    "dev": "nodemon app.js"
+    "dev": "nodemon index.js"
   },
   "author": "Yuvraj",
   "license": "ISC",
